Fix screen shake while resizing
Previously, when the screen was resized during a screen shake, the
game container would be reset to wrong coordinates, because the stored
coordinates (_lastContainerPos) is oblivious to the resize.

This commit moves the responsibility of tracking the original game coordinates
to the Game class to be easily able to react to resize events.